+++ src/js/utilities.js
@@ -107,3 +107,26 @@ export function getScrollParent(element, includeHidden = false) {
     ? $(element.ownerDocument || document)
     : scrollParent;
 }
+
+/**
+ * Add leading zeroes to the given value
+ * @param {number} value - The value to add.
+ * @param {number} [length=1] - The expected value length.
+ * @returns {string} Returns converted value.
+ */
+export function addLeadingZero(value, length = 1) {
+  const str = String(Math.abs(value));
+  let i = str.length;
+  let result = '';
+
+  if (value < 0) {
+    result += '-';
+  }
+
+  while (i < length) {
+    i += 1;
+    result += '0';
+  }
+
+  return result + str;
+}