+++ test/methods/parseDate.js
@@ -1,5 +1,8 @@
 QUnit.test('method.parseDate', function (assert) {
   var $input = window.createInput();
+  var date = $input.datepicker('parseDate', '02/14/2014');
 
-  assert.equal($input.datepicker('parseDate', '02/14/2014').getTime(), new Date(2014, 1, 14).getTime());
+  assert.equal(date.getFullYear(), 2014);
+  assert.equal(date.getMonth(), 1);
+  assert.equal(date.getDate(), 14);
 });